edc1/requirements.txt
backcall==0.1.0
colorama==0.4.1
decorator==4.4.1
Django==2.2.4
django-bootstrap-modal-forms==1.4.4
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
drf-url-filters==0.5.1
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.3
Pygments==2.4.2
pytz==2019.2
sqlparse==0.3.0
Pillow==7.0.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
voluptuous==0.11.7
wcwidth==0.1.7
wcwidth==0.1.7

